[FEATURE] feature/CRAWL-82 Implementation of Grot and Behavior (#18)
* brought in player controller and camera system

* Created Sandbox Scene

-Added rooms prefab to scene
-Implemented Camera prefabs

* fixed camera trigger gizmo

* Renamed Room Root Prefab, and Building out Geometry

-Adjusted camera angles
-Renamed Room prefab to RoomTemplate

* Adjusted Camera angle

-Changed the first camera angle. So the player can see the scene better.

* Created Room 2

-Started working on hallway to room 2
-Added more cameras to the scene

* Added current blockout to Ashley Scene

-Current blockout to Ashley scene for safekeeping

* Added Elevation

-Created a section with stairs
-Created a ramp that goes downward

* Finished creating geometry for Sandbox Level

-Added minor setpieces(Crates and Pillars)
-Adjusted Camera angles
-Built out final two rooms

* Put Sandbox geometry into Ashley Scene

-Copy and pasted blockout into Ashley Scene

* feature/CRAWL-80

Reorganized enemy folder pathing
Basics of Boar Behavior
Basic prototype nodes for behavior for future use, and use in Boar.

* reconnected prefabs in sandbox

* Changed player attack hit box to be trigger.

* feature/CRAWL-80

Tweaking the boar behaviors, ensuring it works with the actual player.
Added the health-based scripts, does take damage, cannot deal damage.

* feature/CRAWL-80

Fixed broken damage nodes so it will be destroyed when hit points are 0.

* feature/CRAWL-80

Minor tweaks for the Grot and Boar behavior, does require a NavMesh for the Grot to work.

* feature/CRAWL-80

Trying to make sure  that the Boar works in the new level. NavMesh not playing nice.

* feature/CRAWL-80

Added some fixes to the Boar, no longer uses Tags to find the player
Still needs to deal damage.

* Blood Particles Added

Added blood particles. They have collision to collide with 3D objects to get more a of realistic feel

* Added boar mesh

Added the mesh for the boar enemy without the rig.

* feature/CRAWL-80

Adding the random wander behavior, does favor the center of the room.

* feature/CRAWL-80

Finalized the boar behavior and worked out some of the kinks regarding attitude.

Also imported the boar head and particle system to the asset.

* Removed duplicate damage system assets.

* Fixed typo in Damage System folder.

* Fixing locations for testing

* removed editor references in FindWithScriptAction

* AnimatedBoar

Created animations for the boar mesh and exported them to unity.

* created animated boar model

* Update Boar.controller

* Setting Up the Boar Animation

* Boar Wandering Behavior Fix

* Boar Presentation Materials

* Grot mesh and animations

Added the grot mesh and animations

* Grot Animation Hookup

Grot Ready for attacking
Animations Set Up
Added a Do Once function to get start location.

* Updated Grot Physicality

Cleaned up Grot Behavior, adjusted the physicality to better suit the required metrics for drag and drop.

* Minor Grot Changes

* Added Bider Model

* Got Grots Done, Ready for QA and UX Tests

* Doubled Grot Size, and sped up

* Fixed Grot Attack Timing

* Closer on timing for Grot Attack

* Fixed the timing with grot's attacks

---------

Co-authored-by: Brevillo <90069242+Brevillo@users.noreply.github.com>
Co-authored-by: ashleycanniff <157764871+ashleycanniff@users.noreply.github.com>
Co-authored-by: Kylie-Lingo <157653967+Kylie-Lingo@users.noreply.github.com>
Co-authored-by: kyoninh <132638948+kyoninh@users.noreply.github.com>
Co-authored-by: Neon-Neko <siggysalmon@gmail.com>
Co-authored-by: MidasTerrarium <151302578+MidasTerrarium@users.noreply.github.com>
Co-authored-by: Travis R. French <travis.french.1991@gmail.com>